refactor(utils): move any defined types to their folder
