+++ galasa-ui/src/components/test-runs/TestRunsTabs.tsx
@@ -18,7 +18,8 @@ import { RESULTS_TABLE_COLUMNS, COLUMNS_IDS, RUN_QUERY_PARAMS, DAY_MS, TABS_IDS,
 import { useQuery } from '@tanstack/react-query';
 import { decodeStateFromUrlParam, encodeStateToUrlParam } from '@/utils/urlEncoder';
 import { TimeFrameValues } from '@/utils/interfaces';
-import { ColumnDefinition, runStructure, sortOrderType } from '@/utils/interfaces';
+import { ColumnDefinition, runStructure } from '@/utils/interfaces';
+import { sortOrderType } from '@/utils/types/common';
 import { Run } from '@/generated/galasaapi';
 
 interface TabConfig {
