+++ galasa-ui/src/utils/types/common.ts
@@ -1,0 +1,8 @@
+/*
+ * Copyright contributors to the Galasa project
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+export type AmPm = 'AM' | 'PM';
+export type sortOrderType = 'asc' | 'desc' | 'none';
